+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix bug where `export` for a qualified import would not work (#1030)
+- Fix a problem where name resolution would fail to flag name errors for
+  non-exported names in incremetal evaluation in REPL (#1031)
 
 ### Security